+++ tests/test_runner.nim
@@ -8,3 +8,4 @@ import test_search
 import test_search_json
 import test_ignore_rules
 import test_file_reader
+import test_file_writer